requirements.txt
psycopg2
python-box
eventlet
pika
pyyaml
